requirements.txt
args==0.1.0
attrs==16.0.0
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.0
certifi==2016.8.31
cffi==1.7.0
click==6.6
clint==0.5.1
colorama==0.3.7
configparser==3.5.0
cryptography==1.5
cssselect==0.9.2
decorator==4.0.10
entrypoints==0.2.2
enum34==1.1.6
Flask==0.11.1
functools32==3.2.3.post2
idna==2.1
ipaddress==1.0.16
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.0
lxml==3.6.4
MarkupSafe==0.23
matplotlib==1.3.0
mistune==0.7.3
MySQL-python==1.2.5
nau-net==0.1
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.2
numpy==1.7.1
parsel==1.0.3
pathlib2==2.1.0
pickleshare==0.7.4
pkginfo==1.3.2
prompt-toolkit==1.0.7
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
PyDispatcher==2.0.5
Pygments==2.1.3
pyOpenSSL==16.1.0
pyparsing==1.5.7
pywin32==220.1
pyzmq==15.4.0
qtconsole==4.2.1
queuelib==1.4.2
requests==2.10.0
requests-toolbelt==0.7.0
Scrapy==1.1.2
scrapyd-client==1.0.1
selenium==2.53.6
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.1
traitlets==4.3.0
twine==1.8.1
Twisted==16.4.0
virtualenv==15.0.1
w3lib==1.15.0
wcwidth==0.1.7
Werkzeug==0.11.11
widgetsnbextension==1.2.6
win-unicode-console==0.5
zope.interface==4.2.0
